+++ frontend/src/components/UI/Stepper.module.css
